devtools/client/memory/test/chrome/test_TreeMap_01.html

<!DOCTYPE HTML>
<html>
<!--
Test that the Tree Map correctly renders onto 2 managed canvases.
-->
<head>
    <meta charset="utf-8">
    <title>Tree component test</title>
    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">

    <script type="application/javascript"
            src="chrome://devtools/content/shared/vendor/d3.js">
    </script>
</head>
<body>
  <!-- Give the container height so that the whole tree is rendered. -->
  <div id="container" style="height: 900px;"></div>

  <pre id="test">
    <script src="head.js" type="application/javascript"></script>
    <script type="application/javascript">
      "use strict";
      window.onload = async function() {
        try {
          const container = document.getElementById("container");

          await renderComponent(TreeMap(TEST_TREE_MAP_PROPS), container);

          const treeMapContainer = container.querySelector(".tree-map-container");
          ok(treeMapContainer, "Component creates a container");

          const canvases = treeMapContainer.querySelectorAll("canvas");
          is(canvases.length, 2, "Creates 2 canvases");
        } catch (e) {
          ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
        } finally {
          SimpleTest.finish();
        }
      };
    </script>
  </pre>
</body>
</html>
